test(storage-resize-images): clean up tests
