+++ _emulator/extensions/firestore-translate-text.env.local
@@ -1,0 +1,6 @@
+LOCATION=europe-west2
+LANGUAGES=en,es
+INPUT_FIELD_NAME=input
+OUTPUT_FIELD_NAME=output
+DO_BACKFILL=false
+COLLECTION_PATH=translations
